Support CPPFLAGS in ALL Makefile's of the system. This to provide support for unusual locations of include files -- like FreeBSD which stores gdbm.h into /usr/local/include, and thus can't find it at autoconfiguration, but which will happily use wrong library when linking -lndbm :-/  Furthermore, this is Standard Thing to do for GNU autoconf.

+++ utils/makedb/dblook.c
@@ -211,7 +211,7 @@ char *argv[];
 	fprintf(stderr,"Key %s not found\n",argv[3]);
 	return 2;
       }
-      printf("siz:%d, dat: %s\n", result.size, (char*)result.data);
+      printf("siz:%ld, dat: %s\n", (long)result.size, (char*)result.data);
     }
 
     (dbfile->close)(dbfile);
@@ -248,7 +248,7 @@ char *argv[];
 	fprintf(stderr,"Key %s not found\n",argv[3]);
 	return 2;
       }
-      printf("siz:%d, dat: %s\n",result.size,(char*)result.data);
+      printf("siz:%ld, dat: %s\n",(long)result.size,(char*)result.data);
     }
 
     (dbfile->close)(dbfile);
